+++ app/src/test/kotlin/com/inout/apiserver/application/word/CreateWordApplicationTest.kt
@@ -2,7 +2,6 @@ package com.inout.apiserver.application.word
 
 import com.inout.apiserver.base.enums.LanguageType
 import com.inout.apiserver.base.service.openai.DictionaryService
-import com.inout.apiserver.base.service.openai.dto.Definition
 import com.inout.apiserver.domain.word.WordFactory
 import com.inout.apiserver.error.BadRequestException
 import com.inout.apiserver.error.ConflictException
@@ -57,7 +56,7 @@ class CreateWordApplicationTest(
         describe("when given word does not exist") {
             describe("when word definition not found") {
                 beforeEach {
-                    doReturn(emptyList<Definition>())
+                    doReturn(emptyList<DictionaryService.Definition>())
                         .whenever(dictionaryService)
                         .fetchWordDefinitions(any(), any(), any())
                 }
@@ -88,7 +87,7 @@ class CreateWordApplicationTest(
                 beforeEach {
                     doReturn(
                         listOf(
-                            Definition(
+                            DictionaryService.Definition(
                                 type = "noun",
                                 definition = "얄리얄리얄랄라",
                                 preContext = "얄라리얄라",
@@ -128,12 +127,12 @@ class CreateWordApplicationTest(
                 beforeEach {
                     val definitions =
                         listOf(
-                            Definition(
+                            DictionaryService.Definition(
                                 type = "noun",
                                 definition = "책",
                                 preContext = "정보를 얻거나 즐거움을 얻기 위해 읽는 인쇄물",
                             ),
-                            Definition(
+                            DictionaryService.Definition(
                                 type = "verb",
                                 definition = "예약하다",
                                 preContext = "특정한 날짜나 시간에 무엇을 하기 위해 미리 자리를 확보하다",
